Print standing penalty names only when patrons are actually blocked.
Add configuration options so the administrator/user can specify what
block_types will cause a patron retrieval to print a standing_penalty
name.  If none of the block types are matched, then the penalty will
not print.  If the patron has only notes or other penalties that don't
match, then ACTIVE or JUVENILE will print to indicate the patron is
valid.  See the comments in issa.xml.example around /issa/patrons
to see how this can be configured.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>