Add COPY_HOLDS_FORCE support to setup.pl.
Check if the COPY_HOLDS_FORCE permission exists in the user's system.
If it does, then offer to use it with issa. After reading about the
consequences, the user is prompted to enter Y or N to indicate whether
or not they want to use the force holds feature with issa.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>